+++ NEWS.md
@@ -1,4 +1,13 @@
-## to be version 0.1.4
+## to be version 0.2
+### Changes
+- ensemble inference, summary, and plot functions now extract the prior list from attribute of the fit objects (previously, the prior_list needed to be passed for each model within the model_list as the priors argument
+
+### Features
+- adding formula interface for fitting and computing marginal likelihood of JAGS models
+- adding vector and factor prior distributions
+
+
+## version 0.1.4
 ### Fixes
 - fixing DOIs in the references file
 - adds marglik argument `inclusion_BF` to deal with over/underflow (Issue #9)